+++ test/function/arithmetic/smallereq.test.js
@@ -35,21 +35,6 @@ describe('smallereq', function() {
     assert.equal(smallereq(false, 2), true);
   });
 
-  it('should compare two complex numbers correctly', function() {
-    assert.equal(smallereq(complex(1,1), complex(1,2)), true);
-    assert.equal(smallereq(complex(1,1), complex(1,1)), true);
-    assert.equal(smallereq(complex(2,4), complex(4,2)), true);
-    assert.equal(smallereq(complex(1,1), complex(2,1)), true);
-    assert.equal(smallereq(complex(1,6), complex(7,1)), true);
-    assert.equal(smallereq(complex(4,1), complex(2,2)), false);
-    assert.equal(smallereq(complex(2,0), 3), true);
-    assert.equal(smallereq(complex(2,0), 2), true);
-    assert.equal(smallereq(complex(2,0), 1), false);
-    assert.equal(smallereq(3, complex(2,0)), false);
-    assert.equal(smallereq(2, complex(2,0)), true);
-    assert.equal(smallereq(1, complex(2,0)), true);
-  });
-
   it('should compare two measures of the same unit correctly', function() {
     assert.equal(smallereq(unit('100cm'), unit('10inch')), false);
     assert.equal(smallereq(unit('99cm'), unit('1m')), true);
@@ -61,7 +46,7 @@ describe('smallereq', function() {
     assert.throws(function () {smallereq(unit('100cm'), 22)});
   });
 
-  it('should perform lexical comprison of two strings', function() {
+  it('should perform lexical comparison of two strings', function() {
     assert.equal(smallereq('0', 0), true);
     assert.equal(smallereq('abd', 'abc'), false);
     assert.equal(smallereq('abc', 'abc'), true);
@@ -71,6 +56,12 @@ describe('smallereq', function() {
   it('should perform element-wise comparison on two matrices', function() {
     assert.deepEqual(smallereq([1,4,6], [3,4,5]), [true, true, false]);
     assert.deepEqual(smallereq([1,4,6], matrix([3,4,5])), matrix([true, true, false]));
+  });
+
+  it('should throw an error when comparing complex numbers', function() {
+    assert.throws(function () {smallereq(complex(1,1), complex(1,2))}, TypeError);
+    assert.throws(function () {smallereq(complex(2,1), 3)}, TypeError);
+    assert.throws(function () {smallereq(3, complex(2,4))}, TypeError);
   });
 
   it('should throw an error with two matrices of different sizes', function () {