+++ test/function/statistics/min.test.js
@@ -15,18 +15,6 @@ describe('min', function() {
     assert.equal(math.min('A', 'C', 'D', 'B'), 'A');
   });
 
-  it('should return the min value for complex values', function() {
-    assert.deepEqual(math.min(math.complex(2,3), math.complex(2,1)), math.complex(2,1));
-    assert.deepEqual(math.min(math.complex(2,3), math.complex(2,5)), math.complex(2,3));
-  });
-
-  it('should return the min value for mixed real and complex values', function() {
-    assert.deepEqual(math.min(math.complex(3,4), 4), 4);
-    assert.deepEqual(math.min(math.complex(3,4), 5), math.complex(3,4));
-    assert.deepEqual(math.min(5, math.complex(3,4)), 5);
-    assert.deepEqual(math.min(math.complex(3,4), 6), math.complex(3,4));
-  });
-
   it('should return the min element from a vector', function() {
     assert.equal(math.min([1,3,5,-5,2]), -5);
   });
@@ -46,6 +34,16 @@ describe('min', function() {
       [ 3, 0,  5],
       [-1, 9, 11]
     ])), math.matrix([-1, 0, 5]));
+  });
+
+  it('should throw an error when called with complex numbers', function() {
+    assert.throws(function () {math.min(math.complex(2,3), math.complex(2,1))}, TypeError);
+    assert.throws(function () {math.min(math.complex(2,3), math.complex(2,5))}, TypeError);
+
+    assert.throws(function () {math.min(math.complex(3,4), 4)}, TypeError);
+    assert.throws(function () {math.min(math.complex(3,4), 5)}, TypeError);
+    assert.throws(function () {math.min(5, math.complex(3,4))}, TypeError);
+    assert.throws(function () {math.min(math.complex(3,4), 6)}, TypeError);
   });
 
   it('should throw an error if called with invalid number of arguments', function() {